chore: mark old sandbox
