+++ package.json
@@ -72,6 +72,7 @@
     "scripts",
     "bun",
     "website",
-    "tests"
+    "tests",
+    "sandbox"
   ]
 }
