+++ packages/sandbox-old/src/components/SkeletonPage.tsx
